+++ Modules~/AbTestFlowModule/Scripts/Runtime/Enums/AbTestValidationSeverity.cs
@@ -1,0 +1,9 @@
+namespace Modules.AbTestFlowModule.Enums
+{
+    public enum AbTestValidationSeverity
+    {
+        Information = 0,
+        Warning = 1,
+        Error = 2
+    }
+}